19 fad development (#21)
These are updates to the loci that are needed to support the new FAD support.

In the CHEM code there were subdirectories for the grid reader and the boundary conditions that were managed in a way that was not sustainable for making a FAD version of the core CHEM rules. The Loci.conf was adjusted so that the make rules could work with directly with the subdirectories to simplify the build process. This is probably something that we will want to include in the makefile changes, but I needed it for the FAD conversion as well.

Classes that include the real type are problematic when trying to make a code that loads two versions of itself. Basically the member functions are not disambiguated by types, so when you load the version of the code where real=FADd into the version of the code where real=double, you can get problems where objects are constructed using a constructor for real type=double, but the destructor is called using the real=FADd or similar types of problems. So these classes need to be disambiguated either by naming the type based on the underlying types used (e.g. Area_FAD and AREA_real) or by making the core type a templated type and using typedef to determine which type is being used. This problem came up specifically for the Area class which I changed to templated type. Note, there might be some other classes that I have overlooked, but right now that was the main problem of this sort that I had to solve.

I needed to update the module loading facilities to make it easier to exclude certain rules from being loaded in the FAD version of the module. I could have done this with #ifdefs in each file, but it is much more straightforward to annotate in the module rule that we don't want to load any rule that computes a certian variable. (for example timestep_finished, that would otherwise be duplicated) The advantage of doing this in the module is that you can be certain that these rules are not being loaded without searching every file to find any offending rules for #ifdef guards.

The module rule was updated slightly to make the name of one of the member functions more readable. The using_nspace method is really specifying what modules this module needs to load into the same namespace as this module. I made it so a user could use the more obvious load_module method to describe the same thing. The using_nspace method still works but is depreciated.

There were some minor changes to diagnostics that I made that helped to track down issues I had when debugging the tangled graph. In addition to this some minor changes the the System makefile that were for similar reasons.