RLF[#23]: Add public accessor for start_time from execute.h (startTim… (#24)
This fixes the stopwatch class so that it can be used by rules inside of a blackbox container.  The main change is making the stop method a const member function so it can be called by rules that input the blackbox container.  This allows for elapsed time to be measured inside of Loci rules to report accurate execution times.  Doxygen comments also added to the class.
---------

Co-authored-by: Raymond L. Fontenot <raymond.fontenot@RLF-RHEL-P3660.cfdrc.com>
Co-authored-by: Ed Luke <edluke@gmail.com>