RLF[28]: added configuration option --no-sub-dir  (#29)
This issue/pull request added the --no-sub-dir feature to the configure
step. The --no-sub-dir feature will disable building
${LOCI_BASE}/Loci-(build info), but instead only use the
--prefix=<install_directory> specified at configure time. The
post-appending of the Loci-(build_info) can be problematic for users on
HPC systems that use module files, as, for example, Loci/CHEM or
Loci/GGFS require ${LOCI_BASE} which is pulled from
--prefix+Loci(build_info). Not specifying --no-sub-dir will keep the
default behavior of the configure/build process. This issue replaces
https://github.com/rlfontenot/loci/issues/25 (due to git issue naming
requirements) and is tied to the makefile issue
https://github.com/rlfontenot/loci/issues/20.

Install.bash updated to support input from configure.
install.txt updated to let the user know the option exists in the build
portion of the guide.
Loci.conf updated to change the install paths.

Co-authored-by: Raymond L. Fontenot <raymond.fontenot@RLF-RHEL-P3660.cfdrc.com>