Code Cleanup to Support clean valgrind runs (#33)
This addresses some valgrind memory debugger warnings that came up in
debugging a memory issue with the online adaption code. This can be
summarized as three major changes: 1) the pack infrastructure used in
message passing was changed to using a memcpy approach instead of
MPI_Pack which had spurious errors. 2) The communication routines used
to manage entity set communications used during existential analysis and
3) Code that implements a generalized global reduction operation that
can work on dynamically sized objects. This is a fairly major
refactoring or these routines. The code has been tested both with
valgrind and using the quicktests for both Loci and chem over a wide
range of processors. More testing is always better.