34 Implement Git submodules to track dependencies (#36)
Closes #34. This pull request implements Git submodule functionality in
order to begin tracking third-party dependencies of the Loci solver,
both mandatory and optional.

The current scope is limited to pinning relatively recent releases of
each library which have been vetted and are known to work with Loci. In
the future this infrastructure could be further developed to support
semi-automated dependency installation and patches to the external
software if ever deemed necessary.

A small update was made to README.md to demonstrate how to use the
submodules to obtain dependencies, and to clarify some of the language.
A more thorough update to both README.md and install.txt is still
needed, but that is outside the scope of this pull request.