Memory bug fix that caused the rule manager to allocate excess memory (#39)
This fix resolves an issue where repeated queries would result in a
slowly increasing baseline memory allocation. This allocation had to do
with the internals of how rules were associated with integer indices
which resulted in non-unique assignment of indices. Ultimately this
meant that rule information was stored many redundant times after
repeated queries. This fixes this bug. The code was tested against the
Loci and chem quicktest suites. The changes should not change the
behavior of Loci with the exception of the memory behavior described
above. Just for safe measure, it should be tested against the CFDRC
codes.