TED[#35]: Adds git-based version and archive Make rules.
The configure and make process were updated to extract and store the
current branch name and a string with key information about the working
tree using git commands. Additionally, a `make tarball` command was
added which generates a tar archive containing all of the git-tracked
files and the version.conf file generated by the previous commands.

Currently the rules to make version.conf and the tarball are constained
in src/Makefile. The tarball command could also be placed in the
Makefile generated by the configure script, but this would require
including ${OBJDIR}/Loci.conf in order to get the git information, which
seems less than ideal.

Note that `git archive` will only include committed changes, while
version.conf will include `-dirty` if uncommitted changes are detected.
We therefore may want to strip that from GIT_INFO as part of the `make
tarball` command.

To close out this issue, we still need to add the strings stored in the
$GIT_INFO and $GIT_BRANCH variables to src/System/Loci_version.cc and
src/lpp/main.cc.