TED[#35]: Pass branch and version strings as macros.
This commit defines LOCI_BRANCH and LOCI_VERSION macros which are
populated at compile time within version.conf. They are currently only
used in lpp/main.cc and System/Loci_version.cc to generate the
revision_name which is used to query version().

Currently the macros are added directly to the DEFINES, but they could
just as easily be targeted to the specific objects which need them.