RLF[#44]: removed #define Map and switched to using Loci::Map (#45)
A long ago (19 years) a version of MPI C++ bindings included Map in the global namespace.  To support compiling with that version of MPI we used a #define to change Map to Loci::Map.  This could cause surprising interactions with other software that happens to use Map as a name.  Since C++ bindings in MPI have since been depreciated (and C++ compilers now uniformly support namespaces) it is unlikely that this hack is no longer needed for portability any more.  Therefore this fix removes this hack so as to avoid any problems interactions with other software.

Co-authored-by: Raymond L. Fontenot <raymond.fontenot@RLF-RHEL-P3660.cfdrc.com>
Co-authored-by: Ed Luke <edluke@gmail.com>