EAL[#47]: Makefile refactoring updates
The most important update here is a change to lpp to include $include
files in make dependency files. Also adds dependency file rule for
".cpp" files that was missing. Also cleanedup and Makefiles to improve
consistency. This needs to be merged to the dev so I can incorporate the
changes to lpp for the documentation facility changes I am working on.