46 lpp documentation interface (#50)
This update integrates javadoc style comments (used in doxygen) with the
Loci preprocessor and Loci framework. It allows javadoc style comments
to be directly injected into the compiled code so that documentation can
be automatically generated via command line arguments and this
documentation can be smartly collated based on Loci specific semantics.
For example, it groups all rules that are parameterized by a parametric
variable together in the documentation. Comments can be added to rules
by simply having a javadoc style comment preceding the rule definition,
no need to use the comment() function. I have updated the FVM.lh to
include javadoc style comments that are parsed by lpp and used to add
documentation to the rule. To see an example of how this works, I have
incorporated the documentation options in the vogcheck utility. So for
example, to see the documentation of all of the rules used to build the
vogcheck utility run the command
vogcheck -doc

If you want to see a more focused set of documentation, you can request
variable specific documentation, for example try
vogcheck -docvar vol

This is a first step to integrating with doxygen, that is so we can use
lpp to generate dox files that will smartly collate documentation in
ways that makes sense for Loci system and rule documentation. This is
just a glimpse into what is possible, but it will be simple to do alot
more in future revisions of this capability.