Ignore library hints from cmake-built HDF5. (#54)
Closes #51. This is a simple workaround for a bug in the configure
script where the format of the "Extra libraries" line in
libhdf5.settings is incompatible with the build process when HDF5 was
built using the cmake infrastructure instead of autotools. The fix is
simply to ignore the line if a semicolon is detected within it. It seems
that these flags are not necessary in general, and could likely be
safely removed.