Fixed line number counter in lpp (#56)
Sometimes lpp was losing count of lines as it processed loci files. This
resulted in error and documentation messages that had incorrect lines.
This was caused by the parser not accounting for lines processed when
processing variables inside of rules that happened to be at the end of
the line. This patch fixes that bug. Tests show that the line numbers
are now correctly counted. Tested with Loci and chem quickTest
regression tests.