Fixes non-ANSI standard compliant code. (#67)
The definition of the MPI types for the MFAD types were allocating
arrays with dynamic sizes. This is not valid ANSI C++ code as the size
needs to be something set at compile time. The code is fixed to use the
constant size instead of an integer that was assigned to this size.