83 fvmmod unit tests (#84)
This is a first pass at adding unit tests for FVMMod capabilities. Right
now it has some basic tests for gradients and limiters (e.g. gradients
of a linear function should be exact, and limiters of linear functions
should be unity (under the right circumstances). More tests are
possible, but these will cover obvious faults. For example, the error in
3-D vector gradients would have been detected by this simple test.